Add files via upload
An entity relationship diagram (ERD) shows the relationships of entity sets stored in a database. An entity in this context is an object, a component of data. An entity set is a collection of similar entities. These entities can have attributes that define its properties. 
By defining the entities, their attributes, and showing the relationships between them, an ER diagram illustrates the logical structure of databases.